changed int to big int for all increments

+++ src/Database/Migrations/2018_04_01_142501_create_dz_routes_table.php
@@ -15,15 +15,15 @@ class CreateDzRoutesTable extends Migration {
         Schema::create('dz_routes', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->bigIncrements('id');
+            $table->increments('id');
 
             $table->string('uri');
             $table->string('method');
 
             $table->string('route_hash');
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('created_by')->references('id')->on($userTable);
             $table->foreign('updated_by')->references('id')->on($userTable);
