+++ src/Database/Migrations/2018_04_12_084320_create_dz_role_assignments_table.php
@@ -18,20 +18,20 @@ class CreateDzRoleAssignmentsTable extends Migration {
             $userTable = LaravelUtility::getUserTable();
             $roleTable = ( new Role() )->getTable();
 
-            $table->bigIncrements('id');
+            $table->increments('id');
 
-            $table->unsignedBigInteger('role_id')->nullable();
+            $table->unsignedInteger('role_id')->nullable();
 
             $table->string('source_type')->nullable();
-            $table->unsignedBigInteger('source_id')->nullable();
+            $table->unsignedInteger('source_id')->nullable();
 
             $table->string('scope')->nullable();
 
             $table->string('target_type')->nullable();
-            $table->unsignedBigInteger('target_id')->nullable();
+            $table->unsignedInteger('target_id')->nullable();
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('role_id')->references('id')->on($roleTable);
